S03P31D105-131 [Edit] Clear error in rank.

See merge request s03-final/s03p31d105!97

+++ gameBack_InGame/src/main/java/com/ssafy/pjt3/service/UserServiceImpl.java
@@ -53,8 +53,8 @@ public class UserServiceImpl implements UserService{
 				if(rankList.get(i).getId() == user_id) break;
 			}
 			
-			double per = index / rankList.size() * 100;
-			
+			double per = (double)index / (double)rankList.size() * 100;
+
 			if(per <= 1) rank = "S";
 			else if(per <= 5) rank = "A";
 			else if(per <= 10) rank = "B";
